requirements.txt
aiodns==2.0.0
aiosmtpd==1.2
asgiref==3.2.10
astroid==2.4.2
atpublic==2.0
cffi==1.14.2
colorama==0.4.3
cryptography==3.0
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
isort==5.4.2
lazy-object-proxy==1.5.1
mccabe==0.6.1
psycopg2==2.8.5
pycares==3.1.1
pycparser==2.20
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
